グラフコンポーネントの名前変更
_をコンポーネントの名前に入れると読み込めない場合があったので_を外した。
